pyproject.toml
[project]
name = "twitter-to-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocron>=2.1",
    "aiosqlite>=0.21.0",
    "discord-py>=2.5.2",
    "pendulum>=3.1.0",
    "pydantic-extra-types>=2.10.4",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
    "twikit>=2.3.3",
    "typing-extensions>=4.13.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.11.11",
]


[tool.mypy]
python_version = "3.10"
ignore_missing_imports = false
warn_unused_ignores = true     # Warns about unneeded # type: ignore comments.
show_error_codes = true

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py312"
respect-gitignore = true

[tool.ruff.lint]
select = [
    "ASYNC",
    "B",
    "BLE",
    "C4",
    "COM",
    "E",
    "ERA",
    "F",
    "FA",
    "FLY",
    "I",
    "INP",
    "ISC",
    "PERF",
    "PIE",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "PT",
    "PTH",
    "PYI",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "SLOT",
    "T10",
    "TCH",
    "TRY",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "PLR09",    # Too many {problem}
    "PLR2004",  # Magic value used in comparison, consider replacing {value} with a constant variable
    "PT011",    # pytest.raises({exception}) is too broad, set the match parameter or use a more specific exception
    "S311",     # Standard pseudo-random generators are not suitable for cryptographic purposes
    "TRY003",   # Avoid specifying long messages outside the exception class
    "TRY301",   # Abstract raise to an inner function
    "UP040",    # Type alias {name} uses TypeAlias annotation instead of the type keyword - mypy does not support this yet
    "ISC001",   # Conflict
    "COM812",   # Conflict
]
task-tags = ["TODO", "FIXME"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
case-sensitive = true
force-single-line = true
lines-after-imports = -1
lines-between-types = 0
order-by-type = false
split-on-trailing-comma = true
